requirements.txt
MetaTrader5==5.0.44
pandas==2.2.2
numpy==1.26.4
python-telegram-bot==20.7
python-dotenv==1.0.1
